ref(bash): Made 'thoughts' function into 'thought' command.
Made improvements to how the code works. Accepts up to two arguments. If two arguments are detected, the first argument determines the file to write to, and the second argument is the text to append to that file. If only one argument, than that is the text to append. There are default variables assigned for the file, directory, and date; All used to build the final message and allow for easy management.

It was rather confusing as I learned how to write bash conditionals and interpolating the arguments into strings and file paths properly.